# perf: EXC-1942: Add `BitVec` create/grow benchmarks